Setup base for javascript
Keeping things simple from the start; will be using jQuery to manage the
validation, and possible eventual communication with 3rd party APIs. If
we feel like React or other framework could help, we can switch it out
later.

Using IIFE to keep things encapsulated and avoid exposing variables in
the global namespace.

Removed javascript call in html since we'll be assigning that in the
javascript - want to keep the js logic and references separate from the
markup.