Add underscore.js
Underscore.js is really nice for adding some helper/utility methods that
you take for granted sometimes using a language like python or ruby.
It's not absolutely vital for a project, and it does come with some
overhead of new developers needing to know another JS tool, but most
of the functions are pretty readable and generally making understanding
the code easier.

If this was going in production right now, clearly the separate
javascript files would want to be concatenated and minified. The entire
gulp build task will need to be returned to later on.