Implement jvfloat plugin
The float label usability pattern allows users to not lose sight of
placeholder text when filling out a field. jvFloat is a jQuery plugin
that implements this pattern.

The jvFloat bower.json lacks a readable "main" property value (jvFloat
tries to list two files, the css and js), and as a result the bower
injection we get "for free" from the wiredep task and respective gulp
plugin no longer works. I've added gulp-inject, a new task, and new
placeholders in index.html to account for this.

Finally, I've created a vendor-overrides directory in `styles`. Ideally,
I'd prefer the styles don't get too coupled with vendor plugins, as we
may decide to remove them - it's even more relevant in this case because
jvFloat adds new DOM elements and corresponding classnames; down the
road, if we remove jvFloat we want to be able to remove the styles
specific to it, easily.