Address autocompletion via google
I decided to use Google's Maps API to implement autocomplete for the
address box. It has the benefit of generally requiring less typing,
provides the user with immediate reapproval that they're entering their
address correctly, and overall, provides a better UX. That said, it
comes with some drawbacks: we now have an external service that we're
depending on - that said, if the API is down, nothing breaks,
it just diminishes the UX and the user has to fill out the form as they
normally would.

Google doesn't seem to have the best support/data for Unit numbers and other "subpremise"
information. This means Address Line 2 can be left empty, which brings
up an interesting UX conundrum since it's not required, but
could definitely be an important piece of information that gets left out.
I think I'll try to address this down the line.

The "Powered by Google" makes things feel a little "cookie cut,"
but it's required by the TOS.

I consider SmartyStreets as another API, but after
surveying both API offerings and documentation, I decided to
go with the better established one. I am curious if SmartyStreets would
be better at dealing with the subpremise info.

Finally, there's much more JS code now - I plan on modularizing it so
it's easier to follow, and will probably refactor it, given time.