Fix enter to select not working for autocomplete
I encountered an issue when keystrokes and enter were used to select the
address from the autocomplete dropdown. This resulted in the full
address being pasted into the Address Line 1 input, and also triggered a
form submission. I fixed both of these.

In the future, I'd like to add tab/enter completion which will select
the first entry without having to click it or use down/up keystrokes.